restructure plugin and fix example
