+++ examples/nats-jetstream-replay/Cargo.toml
@@ -8,38 +8,7 @@ edition = "2024"
 crate-type = ["cdylib"]
 name = "nats_jetstream_replay"
 
-[lints]
-workspace = true
-
 [workspace]
-
-[workspace.lints.rust]
-warnings = "deny"
-# wit-bindgen's `generate!`/`export!` macros emit `unsafe` blocks into this
-# crate, so `unsafe_code` cannot be denied here.
-unused_extern_crates = "warn"
-trivial_numeric_casts = "warn"
-unstable_features = "warn"
-unused_import_braces = "warn"
-unused-lifetimes = "warn"
-unused-macro-rules = "warn"
-
-[workspace.lints.clippy]
-unwrap_used = 'deny'
-expect_used = 'deny'
-panic = 'deny'
-indexing_slicing = 'deny'
-dbg_macro = 'deny'
-join_absolute_paths = 'deny'
-needless_collect = 'deny'
-option_env_unwrap = 'deny'
-panicking_unwrap = 'deny'
-serde_api_misuse = 'deny'
-transmute_ptr_to_ref = 'deny'
-transmute_undefined_repr = 'deny'
-trivially_copy_pass_by_ref = 'deny'
-uninit_vec = 'deny'
-unnecessary_box_returns = 'deny'
 
 [dependencies]
 wit-bindgen = { version = "0.60", features = ["async-spawn", "inter-task-wakeup"] }
